[#3]/refactor: appConfig 와 externalConfig 분리
