+++ src/commons/config/external-config/external-config.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { ExternalConfigServiceService } from './external-config-service.service';
+
+@Module({
+  providers: [ExternalConfigServiceService],
+  exports: [ExternalConfigServiceService],
+})
+export class ExternalConfigModule {}
